ALTAPPS-390: incapsulate path processing inside UrlPathProcessor
