+++ shared/src/commonMain/kotlin/org/hyperskill/app/home/injection/HomeFeatureBuilder.kt
@@ -8,7 +8,7 @@ import org.hyperskill.app.home.presentation.HomeFeature.Action
 import org.hyperskill.app.home.presentation.HomeFeature.Message
 import org.hyperskill.app.home.presentation.HomeFeature.State
 import org.hyperskill.app.home.presentation.HomeReducer
-import org.hyperskill.app.magic_links.domain.interactor.MagicLinksInteractor
+import org.hyperskill.app.magic_links.domain.interactor.UrlPathProcessor
 import org.hyperskill.app.profile.domain.interactor.ProfileInteractor
 import org.hyperskill.app.step.domain.interactor.StepInteractor
 import org.hyperskill.app.streak.domain.interactor.StreakInteractor
@@ -23,7 +23,7 @@ object HomeFeatureBuilder {
         streakInteractor: StreakInteractor,
         profileInteractor: ProfileInteractor,
         stepInteractor: StepInteractor,
-        magicLinksInteractor: MagicLinksInteractor
+        pathProcessor: UrlPathProcessor
     ): Feature<State, Message, Action> {
         val homeReducer = HomeReducer()
         val homeActionDispatcher = HomeActionDispatcher(
@@ -33,7 +33,7 @@ object HomeFeatureBuilder {
             profileInteractor,
             stepInteractor,
             analyticInteractor,
-            magicLinksInteractor
+            pathProcessor
         )
 
         return ReduxFeature(State.Idle, homeReducer)
